make watch util and service
