+++ src/utils/watch.ts
@@ -1,0 +1,8 @@
+import { Mapping } from '../types';
+
+export function isMappingPresent(watchMappings: Mapping[], folderPath: string) {
+    const watchMapping = watchMappings?.find(
+        (mapping) => mapping.folderPath === folderPath
+    );
+    return !!watchMapping;
+}
